refatorando e removendo o componente da Interface
