+++ src/pages/MyRecipes/index.jsx
@@ -3,57 +3,18 @@ import { api } from '../../api'
 import './index.css'
 import * as Tag from './index.js'
 import {
-  Stack,
   Typography,
   Box,
-  useMediaQuery,
-  Tooltip,
-  Link,
-  Rating,
-  Paper,
-  MenuList,
-  MenuItem,
-  ListItemIcon,
-  ListItemText,
-  Divider,
-  Pagination,
-  PaginationItem,
-  Avatar,
-  CircularProgress,
-  AppBar,
-  Tab,
 } from '@mui/material'
-import {
-  Forum as ForumIcon,
-  Favorite as FavoriteIcon,
-  Star as StarIcon,
-  NavigateNext as NavigateNextIcon,
-  Close as CloseIcon,
-  MoreVert as MoreVertIcon,
-  Edit as EditIcon,
-  Delete as DeleteIcon,
-  Details as DetailsIcon,
-  Folder as FolderIcon,
-  Image as ImageIcon,
-  Add as AddIcon,
-  ContentCut,
-  ContentPaste,
-  Cloud,
-  ContentCopy,
-  MoreVert,
-} from '@mui/icons-material'
-import PropTypes from 'prop-types'
-import { green, orange, red } from '@mui/material/colors'
 import { AuthContext } from '../../contexts/AuthContext'
 import { INTERFACE } from '../INTERFACE/index.jsx'
-import { Tabs } from '@mui/base'
 import { useTheme } from '@emotion/react'
 import { AppBarMyrecipes } from './componentes/AppBar'
 import { ProgressMyRecipes } from './componentes/progress'
-
+import { OptionsMyRecipes } from './componentes/options'
+import { CardMyRecipes } from './componentes/cardRecipes'
 function TabPanel(props) {
   const { children, value, index, ...other } = props
-
   return (
     <Typography
       component="div"
@@ -67,36 +28,10 @@ function TabPanel(props) {
     </Typography>
   )
 }
-TabPanel.PropTypes = {
-  children: PropTypes.node,
-  index: PropTypes.number.isRequired,
-  value: PropTypes.number.isRequired,
-}
 
-function a11yProps(index) {
-  return {
-    id: `action-tab-${index}`,
-    'aria-controls': `action-tabpanel-${index}`,
-  }
-}
-
-const fabStyle = {
-  position: 'absolute',
-  bottom: 16,
-  right: 16,
-}
-
-const fabGreenStyle = {
-  color: 'common.white',
-  bgcolor: green[500],
-  '&:hover': {
-    bgcolor: green[600],
-  },
-}
 
 export const MyRecipes = () => {
   const { user } = useContext(AuthContext)
-  const matches = useMediaQuery('(min-width:700px)')
   const theme = useTheme()
   const [value, setValue] = useState(0)
   const [myRecipes, setMyRecipes] = useState([])
@@ -107,11 +42,8 @@ export const MyRecipes = () => {
   const currentRecipes = myRecipes.slice(indexOfFirstRecipe, indexOfLastRecipe)
   const [isLoading, setIsLoading] = useState(true)
   const [progress, setProgress] = useState(0)
-
   const [floatingMenu, setFloatingMenu] = useState(null)
-
   const [handleOptions, setHandleOptions] = useState([])
-
   const handleOptionsOP = (index) => {
     const updatedOptions = {}
     for (let i = 0; i < myRecipes.length; i++) {
@@ -142,15 +74,6 @@ export const MyRecipes = () => {
       }
     }
   }
-
-  const handlePageChange = (event, page) => {
-    setCurrentPage(page)
-  }
-  const noWrap = {
-    overflow: 'hidden',
-    whiteSpace: 'nowrap',
-    textOverflow: 'ellipsis',
-  }
   const handleScroll = () => {
 
   }
@@ -163,7 +86,6 @@ export const MyRecipes = () => {
       try {
         const myRecipesList = await api.myRecipes.get(user.uid)
         setMyRecipes(myRecipesList)
-        // Faça algo com a lista de receitas, como atualizar o estado do componente
       } catch (error) {
         console.error('Erro ao buscar as receitas:', error)
       } finally {
@@ -177,245 +99,42 @@ export const MyRecipes = () => {
   const handleChange = (event, newValue) => {
     setValue(newValue)
   }
-
-  const handleChangeIndex = (index) => {
-    setValue(index)
-  }
   return (
-    <INTERFACE>
-      <>
-        <AppBarMyrecipes
-          handleChange={handleChange}
-          value={value}
-        />
-        {isLoading ? (
-          <>
-            <ProgressMyRecipes progress={progress} />
-          </>
-        ) : (
-          <TabPanel value={value} index={0} dir={theme.direction}>
-            <Tag.Cards>
-              {currentRecipes.map((recipe, index) => {
-                return (
-                  <Tag.Card key={index}>
-                    <Stack width={'100%'}>
-                      <Stack
-                        sx={{
-                          position: 'absolute',
-                          width: '3rem',
-                          height: '3rem',
-                          color: red[900],
-                          bgcolor: 'white',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                        }}
-                      >
-                        <MoreVert onClick={() => handleOptionsOP(index)} />
-                      </Stack>
-                      <Tooltip
-                        sx={{ cursor: 'pointer' }}
-                        title={`Ir para os detalhes  ${recipe?.recipeTitle}`}
-                        followCursor
-                      >
-                        <img className="img" src={recipe?.recipeImage} alt="" />
-                      </Tooltip>
-                      <Stack padding={2} spacing={2}>
-                        <Typography color={'gray'} variant="h6" sx={noWrap}>
-                          <Link
-                            href={`/detailsRecipes/${recipe?.id}`}
-                            color="inherit"
-                            underline="hover"
-                          >
-                            {recipe?.recipeTitle}
-                          </Link>
-                        </Typography>
-                        <Stack direction="row" justifyContent={'space-between'}>
-                          <Stack direction="row" spacing={2}>
-                            <Box color={'#ffa505'}>
-                              <Rating
-                                name={matches ? 'size-medium' : 'size-large'}
-                                disabled
-                                defaultValue={1}
-                              />
-                            </Box>
-                            <Typography variant="p">
-                              {recipe?.starsLikedCounter}
-                            </Typography>
-                            <Box>
-                              {handleOptions[index] && (
-                                <Tag.Options>
-                                  <Tag.PaperOptions
-                                    className={`floating-menu ${floatingMenu === index ? 'open' : ''
-                                      }`}
-                                  >
-                                    <MenuList>
-                                      <MenuItem>
-                                        <ListItemIcon>
-                                          <DetailsIcon fontSize="small" />
-                                        </ListItemIcon>
-                                        <ListItemText>Detalhar</ListItemText>
-                                        <Typography
-                                          variant="body2"
-                                          color="text.secondary"
-                                        >
-                                          Control + V
-                                        </Typography>
-                                      </MenuItem>
-                                      <Link
-                                        href={`/editerecipes/${recipe?.id}`}
-                                      >
-                                        <MenuItem>
-                                          <ListItemIcon>
-                                            <EditIcon fontSize="small" />
-                                          </ListItemIcon>
-                                          <ListItemText>
-                                            Editar minha receita
-                                          </ListItemText>
-                                          <Typography
-                                            variant="body2"
-                                            color="text.secondary"
-                                          >
-                                            Control + E
-                                          </Typography>
-                                        </MenuItem>
-                                      </Link>
-                                      <MenuItem
-                                        onClick={() =>
-                                          deleteRecipe(
-                                            recipe?.id,
-                                            recipe?.recipeTitle
-                                          )
-                                        }
-                                        sx={{
-                                          color: red[500],
-                                        }}
-                                      >
-                                        <ListItemIcon>
-                                          <DeleteIcon
-                                            sx={{
-                                              color: red[500],
-                                            }}
-                                            fontSize="small"
-                                          />
-                                        </ListItemIcon>
-                                        <ListItemText>
-                                          Deletar a receita
-                                        </ListItemText>
-                                        <Typography
-                                          variant="body2"
-                                          color="text.secondary"
-                                        >
-                                          Control + D
-                                        </Typography>
-                                      </MenuItem>
-                                      <Divider />
-                                      <MenuItem
-                                        onClick={() =>
-                                          handleClickOutsideMenu(index)
-                                        }
-                                      >
-                                        <ListItemIcon>
-                                          <CloseIcon fontSize="small" />
-                                        </ListItemIcon>
-                                        <ListItemText>
-                                          Fechar o menu
-                                        </ListItemText>
-                                      </MenuItem>
-                                    </MenuList>
-                                  </Tag.PaperOptions>
-                                </Tag.Options>
-                              )}
-                            </Box>
-                          </Stack>
-                        </Stack>
-                        <Stack
-                          spacing={1}
-                          borderTop={'0.1rem solid #f5f5f5f5'}
-                          direction="row"
-                          width={'100%'}
-                          justifyContent="space-between"
-                        >
-                          <Tag.Author>
-                            <Tag.AuthorImage>
-                              <Avatar
-                                sx={{ bgcolor: orange[500] }}
-                                alt={recipe?.name}
-                                src={user?.photoURL}
-                              />
-                            </Tag.AuthorImage>
-                            <Box
-                              sx={{
-                                display: 'flex',
-                                alignItems: 'flex-start',
-                                justifyContent: 'space-between',
-                                flexDirection: 'column',
-                                height: '100%',
-                                fontSize: '13px',
-                                color: '#565656',
-                              }}
-                              id="info"
-                            >
-                              <Typography sx={noWrap} variant="subtitle1">
-                                {user?.displayName}
-                              </Typography>
-                              <Stack direction="row" spacing={2}>
-                                <Stack direction="row" gap={1}>
-                                  <FavoriteIcon
-                                    fontSize={matches ? 'small' : 'medium'}
-                                  />
-                                  {recipe?.favoritesCounter}
-                                </Stack>
-                                <Stack spacing={2} gap={1} direction="row">
-                                  <ForumIcon
-                                    fontSize={matches ? 'small' : 'medium'}
-                                  />
-                                  {recipe?.commentsCounter}
-                                </Stack>
-                              </Stack>
-                            </Box>
-                          </Tag.Author>
-                          <Tag.ReviewScore
-                            variant={matches ? 'subtitle1' : 'h4'}
-                          >
-                            {recipe?.ranking}
-                          </Tag.ReviewScore>
-                        </Stack>
-                      </Stack>
-                    </Stack>
-                  </Tag.Card>
-                )
-              })}
-            </Tag.Cards>
-          </TabPanel>
-        )}
 
-        {top && (
-          <>
-            {myRecipes.length > itemsPerPage && (
-              <Tag.Pagination
-                spacing={2}
-                sx={{
-                  position: 'relative',
-                  transition: '.3s',
-                  mt: '3',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  width: '100%',
-                }}
-              >
-                <Pagination
-                  count={Math.ceil(recipes.length / itemsPerPage)}
-                  page={currentPage}
-                  onChange={handlePageChange}
-                  renderItem={(item) => <PaginationItem {...item} />}
+    <>
+      <AppBarMyrecipes
+        handleChange={handleChange}
+        value={value}
+      />
+      {isLoading ? (
+        <>
+          <ProgressMyRecipes progress={progress} />
+        </>
+      ) : (
+        <TabPanel value={value} index={0} dir={theme.direction}>
+          <Tag.MenuItemsLinks>
+            {currentRecipes.map((recipe, index) => {
+              return (
+                <CardMyRecipes
+                  options={<>
+                    <OptionsMyRecipes
+                      handleClickOutsideMenu={handleClickOutsideMenu}
+                      floatingMenu={floatingMenu}
+                      index={index}
+                      recipe={recipe}
+                      deleteRecipe={deleteRecipe}
+                    />
+                  </>}
+                  handleOptionsOP={handleOptionsOP}
+                  index={index}
+                  recipe={recipe}
+                  handleOptions={handleOptions}
                 />
-              </Tag.Pagination>
-            )}
-          </>
-        )}
-      </>
-    </INTERFACE>
+              )
+            })}
+          </Tag.MenuItemsLinks>
+        </TabPanel>
+      )}
+    </>
   )
 }
